style(module_name): rename pattern and module names
